[tests] Add broken pipe tests for cache_dump and era_dump
