+++ tests/cache_dump.rs
@@ -115,3 +115,18 @@ fn dump_restore_cycle() -> Result<()> {
 
     Ok(())
 }
+
+//------------------------------------------
+// test no stderr on broken pipe errors
+
+#[test]
+fn no_stderr_on_broken_pipe_xml() -> Result<()> {
+    common::piping::test_no_stderr_on_broken_pipe::<CacheDump>(mk_valid_md, &[])
+}
+
+#[test]
+fn no_stderr_on_broken_fifo_xml() -> Result<()> {
+    common::piping::test_no_stderr_on_broken_fifo::<CacheDump>(mk_valid_md, &[])
+}
+
+//------------------------------------------
