+++ tests/common/era.rs
@@ -14,7 +14,7 @@ use crate::common::test_dir::TestDir;
 
 pub fn mk_valid_xml(td: &mut TestDir) -> Result<PathBuf> {
     let xml = td.mk_path("meta.xml");
-    let mut gen = CleanShutdownMeta::new(128, 256, 32, 4); // bs, nr_blocks, era, nr_wsets
+    let mut gen = CleanShutdownMeta::new(128, 512, 32, 4); // bs, nr_blocks, era, nr_wsets
     write_xml(&xml, &mut gen)?;
     Ok(xml)
 }
@@ -23,7 +23,7 @@ pub fn mk_valid_md(td: &mut TestDir) -> Result<PathBuf> {
     let xml = td.mk_path("meta.xml");
     let md = td.mk_path("meta.bin");
 
-    let mut gen = CleanShutdownMeta::new(128, 256, 32, 4);
+    let mut gen = CleanShutdownMeta::new(128, 512, 32, 4);
     write_xml(&xml, &mut gen)?;
 
     let _file = file_utils::create_sized_file(&md, 4096 * 4096);
